+++ src/parser.h
@@ -115,7 +115,7 @@ typedef struct GLog_ {
   uint8_t piping:1;
   uint8_t log_erridx;
   uint32_t read;                /* lines read/parsed */
-  uint32_t inode;               /* inode of the log */
+  uint64_t inode;               /* inode of the log */
   uint64_t bytes;               /* bytes read on each iteration */
   uint64_t size;                /* original size of log */
   uint64_t length;              /* length read from the log so far */
